requirements.txt
# Pipeline d'enrichissement habitat senior - Dépendances

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Web scraping
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.0.0

# AI/LLM
anthropic>=0.21.0
openai>=1.0.0
tavily-python>=0.3.0
groq

# Geocoding
geopy>=2.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Caching (optionnel)
redis>=5.0.0

# Interface
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
